move types into file. import type from brandsdk
