+++ packages/template-block/src/components/TemplatePreview.tsx
@@ -1,8 +1,8 @@
 /* (c) Copyright Frontify Ltd., all rights reserved. */
 
-import { AppBridgeBlock, Template, useBlockAssets } from '@frontify/app-bridge';
+import { useBlockAssets } from '@frontify/app-bridge';
 import { IconPlus24 } from '@frontify/fondue';
-import { PreviewHeightType, Settings, previewHeightValues, textPositioningToFlexDirection } from '../types';
+import { PreviewHeightType, previewHeightValues, textPositioningToFlexDirection } from '../types';
 import {
     BlockInjectButton,
     getBackgroundColorStyles,
@@ -10,14 +10,7 @@ import {
     toRgbaString,
 } from '@frontify/guideline-blocks-settings';
 import { PreviewImage } from './PreviewImage';
-
-export type TemplatePreviewProps = {
-    appBridge: AppBridgeBlock;
-    blockSettings: Settings;
-    template: Template | null;
-    updateBlockSettings: (newSettings: Partial<Settings>) => Promise<void>;
-    onOpenTemplateChooser: () => void;
-};
+import { TemplatePreviewProps } from './types';
 
 export const TemplatePreview = ({
     appBridge,
