+++ packages/template-block/src/components/types.ts
@@ -1,11 +1,44 @@
 /* (c) Copyright Frontify Ltd., all rights reserved. */
 
-import { MenuItemStyle } from '@frontify/fondue';
-import { JSX } from 'react';
+import { AppBridgeBlock, Template } from '@frontify/app-bridge';
+import { Settings } from '../types';
 
-export type FlyoutToolbarItem = {
+export type AlertErrorProps = {
+    errorMessage: string;
+};
+
+export type CustomButtonProps = {
+    blockSettings: Settings;
+    isEditing: boolean;
+    isDisabled: boolean;
+    updateBlockSettings: (newSettings: Partial<Settings>) => Promise<void>;
+    handleNewPublication: () => void;
+};
+
+export type PreviewImageProps = {
+    appBridge: AppBridgeBlock;
+    blockSettings: Settings;
+    template: Template | null;
+    updateBlockSettings: (newSettings: Partial<Settings>) => Promise<void>;
+    onOpenTemplateChooser: () => void;
+};
+
+export type TemplatePreviewProps = {
+    appBridge: AppBridgeBlock;
+    blockSettings: Settings;
+    template: Template | null;
+    updateBlockSettings: (newSettings: Partial<Settings>) => Promise<void>;
+    onOpenTemplateChooser: () => void;
+};
+
+export type TemplateTextProps = {
+    appBridge: AppBridgeBlock;
+    blockSettings: Settings;
     title: string;
-    onClick: () => void;
-    icon: JSX.Element;
-    style?: MenuItemStyle;
+    description: string;
+    pageCount: number | undefined;
+    isEditing: boolean;
+    key: string;
+    setTitle: (newTitle: string) => void;
+    setDescription: (newDescription: string) => void;
 };
